WinjaCTF_NullconGoa2022/UNOOSA/requirements.txt
Flask==2.1.3
PyJWT==2.4.0
